requirements.txt
typer==0.4.0
GitPython==3.1.24
inquirerpy==0.3.0